API for Simulation

+++ model/user.go
@@ -2,7 +2,7 @@ package model
 
 type User struct {
 	Username string `db:"username"`
-	UserId   int    `db:"user_id"`
+	UserId   int64  `db:"user_id"`
 	Email    string `db:"email"`
 	PwHash   string `db:"pw_hash"`
 }
